+++ app/src/main/java/com/dam/juanmaseatitapp/ViewHolder/CartAdapter.java
@@ -42,7 +42,7 @@ public class CartAdapter extends RecyclerView.Adapter<CartViewHolder> {
 
         Locale locale = new Locale("es", "ES");
         NumberFormat fmt = NumberFormat.getCurrencyInstance(locale);
-        int price = (Integer.parseInt(listData.get(position).getPrice())) * (Integer.parseInt(listData.get(position).getQuantity()));
+        double price = (Double.parseDouble(listData.get(position).getPrice())) * (Integer.parseInt(listData.get(position).getQuantity()));
 
         holder.txt_price.setText(fmt.format(price));
         holder.txt_cart_name.setText(listData.get(position).getProductName());
